% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.r
\name{confidence}
\alias{confidence}
\title{Calculate confidence of the predictive distribution model}
\usage{
confidence(observations, predictions, thresholds, type)
}
\arguments{
\item{observations}{# TODO}

\item{predictions}{# TODO}

\item{thresholds}{# TODO}

\item{type}{# TODO}
}
\value{
# TODO
}
\description{
# TODO
}
\examples{
set.seed(12345)
# TODO
\dontrun{
# TODO
}
}
\seealso{
\code{\link{thresholds}} for calculating the two thresholds,
  \code{\link{consistence}} for calculating consistence
}
