% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistence.r
\name{consistence}
\alias{consistence}
\title{Calculate consistence of the predictive distribution model}
\usage{
consistence(conf_train, conf_eval)
}
\arguments{
\item{conf_train}{# TODO}

\item{conf_eval}{# TODO}
}
\value{
# TODO
}
\description{
# TODO
}
\examples{
set.seed(12345)
# TODO
\dontrun{
# TODO
}
}
\seealso{
\code{\link{thresholds}} for calculating the two thresholds,
  \code{\link{confidence}} for calculating confidence
}
