% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.r
\name{thresholds}
\alias{thresholds}
\title{Thresholds needed to create the extended confusion matrix}
\usage{
thresholds(observations, predictions)
}
\arguments{
\item{observations}{An integer or logical vector containing the binary
observations where presences are encoded as \code{1}s/\code{TRUE}s and
absences as \code{0}s/\code{FALSE}s.}

\item{predictions}{A numeric vector containing the predicted probabilities of
occurrence typically within the \code{[0, 1]} interval.
\code{length(predictions)} should be equal to \code{length(observations)}
and the order of the elements should match.}
}
\value{
A named numeric vector of length 2. The first element
  ('\code{threshold1}') is the mean of probabilities predicted to the absence
  locations distinguishing certain negatives (certain absences) from
  uncertain predictions. The second element ('\code{threshold2}') is the mean
  of probabilities predicted to the presence locations distinguishing certain
  positives (certain presences) from uncertain predictions.
}
\description{
Calculate the two thresholds distinguishing certain negatives/positives from
uncertain predictions. The thresholds are needed to create the extended
confusion matrix and are further used for confidence calculation.
}
\note{
Should be called using the full dataset containing both training and
  evaluation locations.
}
\examples{
set.seed(12345)
thresholds(observations = c(rep(x = TRUE, times = 50), rep(x = FALSE, times = 50)), predictions = c(runif(n = 50, min = 0.3, max = 1), runif(n = 50, min = 0, max = 0.7)))
thresholds(observations = c(rep(x = 0L, times = 300), rep(x = 1L, times = 100)), predictions = c(runif(n = 300, min = 0, max = 0.6), runif(n = 100, min = 0.4, max = 1)))
\dontrun{
thresholds(observations = c(rep(x = 0, times = 300), rep(x = 1, times = 100)), predictions = c(runif(n = 300, min = 0, max = 0.6), runif(n = 100, min = 0.4, max = 1))) # throw a warning
thresholds(observations = c(FALSE, FALSE, TRUE, TRUE), predictions = c(0.2, 0.4, 0.7, 1.1)) # throw a warning
}
}
\seealso{
\code{\link{confidence}} for calculating confidence,
  \code{\link{consistence}} for calculating consistence
}
